FINAL/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
duckdb
matplotlib
numpy
requests
scipy
seaborn
python-multipart
lxml
html5lib
beautifulsoup4streamlit
